# Significant performance gain through improved edge indexing