@jonasjancarik 2017-09-26
This really is a dramatic improvement in terms of performance, thanks for submitting the pull request!